policy-states-collector/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
aiohttp~=3.8.4
aiolimiter~=1.0.0
azure-data-tables~=12.4.2
azure-functions~=1.13.2
azure-identity~=1.12.0
azure-mgmt-policyinsights~=1.0.0
azure-core~=1.26.4
azure-mgmt-resource~=22.0.0
azure-monitor-ingestion~=1.0.1
requests~=2.28.2
pytest
